feat(remove): Remove animation js for fade out
Fading a turbo frame in is very simple to do. Fading out is more complicated.

For now I am removing the fade out JS and reverting to the fade in only.

I may attempt to do a fade out stimulus controller on my own though.

I didn't want to permanently use that copied code,
I just wanted to see if it was possible to achieve the concept.