fix: javascript dark mode no longer breaks using card button
Not quite sure why yet, but it seems the JS is unable to figure out
where the dark mode button has gone during turbo transitions
made from card links unless I specifically include the data-turbo-frame
attribute. This has no bearing on how the frame works, as it defaults to
the content-slider. but this makes the javascript happy and work!